+++ packages/build-utils/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @vercel/build-utils
 
+## 10.1.0
+
+### Minor Changes
+
+- [envs] add `VERCEL_TARGET_ENV` to prefixed env vars ([#13069](https://github.com/vercel/vercel/pull/13069))
+
+## 10.0.1
+
+### Patch Changes
+
+- [build-utils] Move `runNpmInstallSema` closer to where it's used ([#13061](https://github.com/vercel/vercel/pull/13061))
+
+- [build-utils] extract getInstallCommandForPackageManager to module scope ([#13058](https://github.com/vercel/vercel/pull/13058))
+
+- [build-utils] remove redundant sema release ([#13059](https://github.com/vercel/vercel/pull/13059))
+
+- Revert support pnpm 10 ([#13064](https://github.com/vercel/vercel/pull/13064))
+
+## 10.0.0
+
+### Major Changes
+
+- Detect v9 pnpm lockfiles as pnpm 10 generated ([#12852](https://github.com/vercel/vercel/pull/12852))
+
+## 9.3.1
+
+### Patch Changes
+
+- [build-utils] extract install command specific logic into helper ([#13049](https://github.com/vercel/vercel/pull/13049))
+
+- [build-utils] remove dead node 16 check ([#13047](https://github.com/vercel/vercel/pull/13047))
+
+## 9.3.0
+
+### Minor Changes
+
+- Support process tracing ([#12894](https://github.com/vercel/vercel/pull/12894))
+
 ## 9.2.1
 
 ### Patch Changes